backend/requirements.txt
Flask>=3.0,<3.1
Flask-CORS>=4.0,<4.1
blinker>=1.7,<1.8
click>=8.1,<8.2
itsdangerous>=2.1,<2.2
Jinja2>=3.1,<3.2
MarkupSafe>=2.1,<2.2
Werkzeug>=3.0,<3.1
PyJWT>=2.8,<2.9
cryptography>=42.0,<43.0 